pbxproj 파일 수정 사항 반영
