--- SavePets/SavePets/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  SavePets
-//
-//  Created by Haeseok Lee on 2021/04/11.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
